requirements.txt
annotated-types==0.6.0
anyio==4.3.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
config==0.5.1
cryptography==42.0.5
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.110.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.7
jwt==1.3.1
priority==2.0.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
pymongo==4.7.0
python-config==0.1.2
python-decouple==3.8
sniffio==1.3.1
starlette==0.37.2
taskgroup==0.0.0a4
tomli==2.0.1
typing_extensions==4.11.0
uvicorn==0.29.0
wsproto==1.2.0
